move isValidClosurePropertyName to type_translator
I want to use this function from within type_translator, and in its
current location if I import it I get a cycle.
